backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
openai
python-dotenv
python-multipart
pydantic
